MAINT: random: Fix a few compiler warnings.
Fixes:

    gcc: numpy/random/src/xoshiro256/xoshiro256.c
    numpy/random/src/xoshiro256/xoshiro256.c:39:16: warning: comparison of integers of different signs: 'int' and 'unsigned long' [-Wsign-compare]
            for (i = 0; i < sizeof JUMP / sizeof *JUMP; i++)
                        ~ ^ ~~~~~~~~~~~~~~~~~~~~~~~~~~

    gcc: numpy/random/src/xoshiro512/xoshiro512.c
    numpy/random/src/xoshiro512/xoshiro512.c:43:17: warning: comparison of integers of different signs: 'int' and 'unsigned long' [-Wsign-compare]
      for (i = 0; i < sizeof JUMP / sizeof *JUMP; i++)
                  ~ ^ ~~~~~~~~~~~~~~~~~~~~~~~~~~
    numpy/random/src/xoshiro512/xoshiro512.c:46:23: warning: comparison of integers of different signs: 'int' and 'unsigned long' [-Wsign-compare]
            for (w = 0; w < sizeof s_placeholder / sizeof *s_placeholder; w++)
                        ~ ^ ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

    gcc: numpy/random/src/distributions/distributions.c
    numpy/random/src/distributions/distributions.c:185:14: warning: comparison of integers of different signs: 'int64_t' (aka 'long long') and 'const uint64_t' (aka 'const unsigned long long') [-Wsign-compare]
        if (rabs < ki_double[idx])
            ~~~~ ^ ~~~~~~~~~~~~~~
    numpy/random/src/distributions/distributions.c:230:14: warning: comparison of integers of different signs: 'int32_t' (aka 'int') and 'const uint32_t' (aka 'const unsigned int') [-Wsign-compare]
        if (rabs < ki_float[idx])
            ~~~~ ^ ~~~~~~~~~~~~~

+++ numpy/random/src/xoshiro512/xoshiro512.c
@@ -6,6 +6,7 @@ worldwide. This software is distributed without any warranty.
 
 See <http://creativecommons.org/publicdomain/zero/1.0/>. */
 
+#include <stddef.h>
 #include "xoshiro512.h"
 
 /* This is xoshiro512** 1.0, an all-purpose, rock-solid generator. It has
@@ -32,7 +33,7 @@ static uint64_t s_placeholder[8];
 
 void xoshiro512_jump(xoshiro512_state *state) {
 
-  int i, b, w;
+  size_t i, b, w;
   static const uint64_t JUMP[] = {0x33ed89b6e7a353f9, 0x760083d7955323be,
                                   0x2837f2fbb5f22fae, 0x4b8c5674d309511c,
                                   0xb11ac47a7ba28c25, 0xf1be7667092bcc1c,
